Fixes to esriLoader to enabled passing an array of strings to it
The current implementation allows an array of strings to be passed to
the loader, but $q Resolves only support 1 parameter.  Therefore, in
your .then() function you would only ever get the first module in your
list.  The loader will now take a callback function that will be passed
all of the requested modules.  The original implementation is also
intact so as not to break an application if someone decides to update to
this version.