more refinements to the map directive lifecycle
update two-way bound scope properties (center, zoom) on load *only*
if the map is already loaded (i.e. from a web map)

call extent change handler even if already in update cycle (fixes a bug
where it was not being called when clicking the "San Francisco" and "New
York" buttons in the center and zoom example.